+++ www/src/lib/components/Favorites.svelte
